Moved `InMemoryFeedStore` to production target in a dedicated file
