+++ Tests/FeedStoreChallengeTests.swift
@@ -5,37 +5,6 @@
 import XCTest
 import FeedStoreChallenge
 
-class InMemoryFeedStore: FeedStore {
-	
-	private struct Cache {
-		let feed: [LocalFeedImage]
-		let timestamp: Date
-	}
-	
-	private var cache: Cache?
-	
-	func deleteCachedFeed(completion: @escaping DeletionCompletion) {
-		cache = nil
-		
-		completion(nil)
-	}
-	
-	func insert(_ feed: [LocalFeedImage], timestamp: Date, completion: @escaping InsertionCompletion) {
-		cache = Cache(feed: feed, timestamp: timestamp)
-		
-		completion(nil)
-	}
-	
-	func retrieve(completion: @escaping RetrievalCompletion) {
-		guard let cache = cache else {
-			completion(.empty)
-			return
-		}
-		
-		completion(.found(feed: cache.feed, timestamp: cache.timestamp))
-	}
-}
-
 class FeedStoreChallengeTests: XCTestCase, FeedStoreSpecs {
 	
 	func test_retrieve_deliversEmptyOnEmptyCache() {
